+++ lib/recording_studio_stripe/update_price.rb
@@ -2,6 +2,8 @@
 
 module RecordingStudioStripe
   class UpdatePrice
+    include StripeRequest
+
     def self.call(price:, metadata:)
       new(price: price, metadata: metadata).call
     end
@@ -12,7 +14,7 @@ module RecordingStudioStripe
     end
 
     def call
-      @price.metadata = (@price.metadata || {}).stringify_keys.merge(@metadata)
+      @price.metadata = merge_metadata(@price.metadata, @metadata)
       update_stripe
       @price.save!
       @price
@@ -24,26 +26,23 @@ module RecordingStudioStripe
       return if RecordingStudioStripe.configuration.local_mode?
 
       stripe_price = Client.current.v1.prices.retrieve(@price.stripe_id)
-      existing = stringify(stripe_get(stripe_price, :metadata))
+      existing = stringify_metadata(stripe_get(stripe_price, :metadata))
       Client.current.v1.prices.update(
         @price.stripe_id,
-        { metadata: existing.merge(@metadata) }
+        { metadata: merge_metadata(existing, @metadata) }
       )
     end
 
-    def stringify(metadata)
-      return {} if metadata.blank?
-
-      metadata.to_h.stringify_keys
-    end
-
-    def stripe_get(object, key)
-      return if object.nil?
-      return object[key] || object[key.to_s] || object[key.to_sym] if object.is_a?(Hash)
-
-      object.public_send(key)
-    rescue NoMethodError
-      object[key] || object[key.to_s] if object.respond_to?(:[])
+    def merge_metadata(current, incoming)
+      data = stringify_metadata(current)
+      incoming.each do |key, value|
+        if value.blank?
+          data.delete(key)
+        else
+          data[key] = value.to_s
+        end
+      end
+      data
     end
   end
 end